requirements.txt
werkzeug==0.16.1
setuptools==33.1.1
wheel==0.33.6
flask==1.0.2
flask-session==0.3.2
flask-helper==0.19
cache==1.0.3
flask-caching==1.8.0
gunicorn==19.9
gevent==1.3.7
gevent-websocket
psutil==5.6.6
pillow==6.2.2
chardet==3.0.4
flask-sqlalchemy==2.3.2
cryptography==3.3.2
requests==2.20.0
ConfigParser==3.5.0
MySQL-python==1.2.5
mysql-connector-python
bidict==0.18.3
python-socketio==4.0.0
flask-socketio==3.0.1
python-engineio==3.9.0
# pytest-shutil
